// src/apps/hotspot/enums/hotspot.enum.ts
export enum HOTSPOT_EVENT_TYPES {
  CLIQUE_UPDATED = 'CLIQUE_UPDATED',
  CLIQUE_DELETED = 'CLIQUE_DELETED',
  NEW_CLIQUE = 'NEW_CLIQUE',
  CLIQUE_MEMBER_REQUEST = 'CLIQUE_MEMBER_REQUEST',
  CLIQUE_MEMBER_CANCEL = 'CLIQUE_MEMBER_CANCEL',
  CLIQUE_MEMBER_ACCEPT = 'CLIQUE_MEMBER_ACCEPT',
  CLIQUE_MEMBER_DECLINE = 'CLIQUE_MEMBER_DECLINE',
  CLIQUE_MEMBER_ADDED = 'CLIQUE_MEMBER_ADDED',
  CLIQUE_MEMBER_REMOVED = 'CLIQUE_MEMBER_REMOVED',
  CLIQUE_MEMBER_LEFT = 'CLIQUE_MEMBER_LEFT',
  NEW_CLIQUE_INTEREST = 'NEW_CLIQUE_INTEREST',
  CLIQUE_UNINTEREST = 'CLIQUE_UNINTEREST',
  NEW_RESOURCE_INTEREST = 'NEW_RESOURCE_INTEREST',
  RESOURCE_UNINTEREST = 'RESOURCE_UNINTEREST',
}

export enum HOTSPOT_PREMIUM_SUBSCRIPTIONS {
  
}

export enum HOTSPOT_NOTIFICATION_TARGET_TYPES {
  CLIQUE = 'CLIQUE',
  RESOURCE = 'RESOURCE',
}

export enum HOTSPOT_SUBSCRIPTION_TARGET_ACTIONS {
  
}

export enum HOTSPOT_SUBSCRIPTION_TARGET_ACTIONS_INFO {
  
}

export enum HOTSPOT_RESOURCE_TYPES {
  BOOK = 'BOOK',
  VIDEO = 'VIDEO',
  EVENT = 'EVENT',
  ARTICLE = 'ARTICLE',
  WEBSITE = 'WEBSITE',
  PERSON = 'PERSON',
  OTHER = 'OTHER',
}

export enum NoticeVisibility {
  PUBLIC = 'PUBLIC',
  PRIVATE = 'PRIVATE',
}